requirements.txt
# Core frameworks
fastapi==0.117.1
uvicorn==0.24.0
streamlit==1.49.0
Django==5.2.6
starlette==0.48.0

# Data Science & AI
numpy==1.24.3
pandas==2.3.2
polars==1.33.1
scipy==1.15.3
matplotlib==3.10.6
plotly==5.15.0
plotly-express==0.4.1
altair==5.5.0
torch==2.8.0
torchvision==0.23.0
ultralytics==8.3.203
ultralytics-thop==2.0.17
sympy==1.14.0
networkx==3.4.2
pyarrow==21.0.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.0.1

# Web & API
requests==2.31.0
httpx==0.24.1
websockets==12.0
websocket-client==1.6.1
anyio==4.11.0
h11==0.14.0
httpcore==0.17.3
sniffio==1.3.1

# Database
psycopg2-binary==2.9.7
sqlparse==0.5.3

# Supabase ecosystem
supabase==2.0.2
gotrue==1.3.1
postgrest==0.13.2
realtime==1.0.6
storage3==0.6.1
supafunc==0.3.3

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.7.1
pydantic-settings==2.11.0
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0
rootutils==0.1.2
utils==0.1.0
picker==2.3.0

# Serialization / Schema
PyYAML==6.0.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1

# Streamlit ecosystem
pydeck==0.9.1
streamlit-autorefresh==1.0.1
streamlit-js-eval==0.1.7

# Dev / Monitoring
watchdog==6.0.0
colorama==0.4.6
click==8.3.0
blinker==1.9.0

# Version control / Git
GitPython==3.1.45
gitdb==4.0.12

# Misc
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
toml==0.10.2
attrs==25.3.0
deprecation==2.1.0
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
Jinja2==3.1.6
MarkupSafe==3.0.2
tenacity==9.1.2
contourpy==1.3.2
cycler==0.12.1
kiwisolver==1.4.9
mpmath==1.3.0
narwhals==2.5.0
psutil==7.1.0
six==1.17.0
smmap==5.0.2
StrEnum==0.4.15
tornado==6.5.2
tzdata==2025.2
python-dateutil==2.9.0.post0
exceptiongroup==1.3.0
asgiref==3.9.2
cachetools==6.2.0
